pyproject.toml
[project]
name = "golf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "gunicorn>=23.0.0",
    "ipython>=8.32.0",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "rq>=2.1.0",
    "rq-scheduler>=0.14.0",
    "sentry-sdk>=2.20.0",
]
